requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
attrs==25.3.0
av==14.4.0
cffi==1.17.1
cryptography==45.0.7
dnspython==2.7.0
frozenlist==1.7.0
google-crc32c==1.7.1
idna==3.10
ifaddr==0.2.0
multidict==6.6.4
propcache==0.3.2
pycparser==2.22
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
PyYAML==6.0.2
typing_extensions==4.15.0
yarl==1.20.1
